Reverse url/text and title.
This is an attempt at mild social engineering to make Reddit better by making
titles more intentional.

Putting the title after the body of self-post should make people consider the
title as a summary instead of as a first line in a message.  After they type
the text, they've fully explored what they want to say, and are in a better
position to give a name to what they had in mind.

For instance, "Hey /r/android, I got my first Android phone the other day
(Nexus S) and I have a question" should be "How can I send SMS longer than 160
characters?", but the user treated the first field as the place to put the
first sentence of his "letter" to us, none of which is relevant.