Update printable.html.
Remove `<!--IE6sux-->` comments.

Do enough users use IE6 to send these bytes to every user?
I use the latest version of Chrome, and when I visit the home page, this comment is sent about 70 times.
That means 910 bytes are sent, just to show this comment.
At the time of writing, last month, this comment was responsible for 4535.956 GB of data sent to users. That is, over 4 TB in a month. All for a comment to say IE6 sucks.

Since Reddit is aiming to become a profitable entity, I would recommend getting rid of this comment, and forcing IE6 users to upgrade. It will be about time.