Fix NameError when populating db.
- Included `import random` at the top of r2/r2/lib/cache.py
- Removed `import random`'s that appear in r2/r2/lib/cache.py function definitions.

There are several usages of `random.random()` in function definitions that do not first `import random`. When populating the database for the first time it throws a NameError:

>>> from r2.models import populatedb
>>> populatedb.populate()

NameError: global name 'random' is not defined

By moving `import random` to the global scope of the file this error is removed.