Added SVG support to custom images
SVG support for subreddit stylesheets allows reddit to move forward
into the future of high-resolution displays, because scalability is key.

In making this change, there are a number of security issues that need
to be considered with relation to SVG and XML files that may be
uploaded to a server, especially one the size of reddit. As such, this
change employs a whitelist of allowable tags and attributes (much like
the CSS validator), and takes all necessary steps to make the SVGs are
secure as possible.

The program, when tested on [my own server](http://50.116.12.143/),
correctly sanitised each and every vulnerability on the [HTML5Sec
page](https://html5sec.org/#svg) that was listed under SVG _and_ XML.
That being said, I would be happy to aid any additional security
testing in any way that I can.