Don't send Cache-Control: no-cache, just do private and max-age=0
Now that reddit has fully switched to https I have noticed the annoyance
that each time I click a link, then click the back button reddit will
refresh the page instead of using the cache like a normal website does
when you use the back button.

This is because reddit is sending a Cache-Control: no-cache header.
A long time ago browsers decided that if an https page sends this it
shouldn't use a cached copy on the back button because of some compatibility
issues with some bank websites: https://bugzilla.mozilla.org/show_bug.cgi?id=567365

For this reason other big sites like news.ycombinator.com use a
Cache-Control: private, max-age=0 so proxies don't cache the response,
browsers don't cache it on a page refresh, but the browser is still allowed
to cache the page for the back button: http://www.w3.org/Protocols/HTTP/Issues/cache-private.html

I think it's good if reddit also does this. This would mean faster serving
for people that don't open links in tabs but instead use the back button.
It would probably also lower the amount of requests reddit gets by a bit.