Make /about/nsfw listing
Make a listing for nsfw posts for the use of mods so that they can see it when needed, easily.

The listing will "not exist" for NSFW-only subs.

If a subreddit that is NSFW, becomes un-NSFW, unfortunately, this listing will not be updated, due to the fact that the posts themeselves don't stay NSFW.