Move generation of S attribute to builder.py
Moves the generation of the S attribute to builder.py, previously in link.py's Comments' `add_props`.

This change makes the S attribute consistent with all the others, and will make it show up in all the places that it is supposed to, such as the inbox, which was the reason why [this bug report was originally filed on /r/bugs](https://redd.it/3oi9n6)

The order of the attributes are kept as they were.