Hotfix: Mod 2 non-mail mod modmail appears in the recipient inbox.
Previously, if a mod sent a non mod a message via the subreddit (ex, so that a single thing can be discussed without that mod having full access to the modmail), it would not appear in the recipient inbox because the check did not care for or for not the mail perm, as reported [here](http://redd.it/3243tc). By changing it to an `is_moderator_with_perms` check, this issue is resolved.