Show moderator invite permissions on /about/moderators
So that users know what they are getting into, display the invited user's perms on /about/moderators when they are going to accept / deny the invite.

What it looks like: https://i.imgur.com/Sj32W5m.png

Of course, if the perms are different, there will be a different list. If the user is invited with no perms, there is no list, and if the user is invited with no / all perms, the displayed string is slightly different for the sake of grammar.