Allow posts/comments by Automoderator to be edited by admins and mods.
Allows posts/comments by Automoderator to be edited by mods if the comment/post is within their sub and they have the "posts" permission. Also allows admin editing.

It can be argued that `getattr(Account.automoderator_user(), '_id', None)` can be slow. However, that can be changed to `g.automoderator_id` if `automoderator_id` is set in the config file. Furthermore, if that is done, `Account.automoderator_user()` can be changed to lookup the account based off of that instead of off of name, to minimize on redundant variables in the config.