Add num_new_reports to the API and change num_reports to be the total_reports
This adds the new reports number to the API as num_new_reports and respects the previous behavior changes of reports on the front end.

Someone may want to check this out locally, since while I'm pretty sure this is correct and I'm having no error, the discrepancy for how the Comment and Link templates render this info is strange to me.