Bug fixes and improvements to proxy support
1.  Add support for non-local (ie not on 127.0.0.1 or 10.0.0.0/8)
    trusted proxies by adding new INI keyword `trusted_proxy_ips`
    (comma separated list of IPs)

2.  Bug fixes in `X-Forwarded-For` processing:
        a. remove whitespace around extracted IP which causes the
           memcached exception 'bad key' (see [1] for an example
           stacktrace)

        b. extract correct IP when multiple IPs are present.  The de
           facto spec[2] says that each proxy *appends* its peer's
           IP, which means the original requestor is the first IP,
           not the last.

[1] http://pastebin.com/1QQiRsgs

In this instance, the request is https terminated by nginx and then
routed to Reddit via Varnish cache (instead of HAProxy) which
automatically appends the remote IP per the de facto spec[2] for the
X-Forwarded-For header.

Because there are effectively two proxies involved, the XFF header
contains two IP addresses separated by a comma and a space.  Having
instrumented lib/cache.py:634, argument `keys` contained (note the
leading whitespace) `[' 127.0.0.1']`.  See the stack trace for the
original contents of the XFF header.

[2] X-Forwarded-For is an uncodifed, de facto standard (now replaced
    by RFC 7239), however there are a number of reputable resources
    that show how it is supposed to work, for example:

http://docs.aws.amazon.com/ElasticLoadBalancing/latest/DeveloperGuide/x-forwarded-headers.html#x-forwarded-for
https://www.citrix.com/blogs/2012/09/27/multiple-xff-ips/