Fix format of subreddit titles in system messages
Subreddit titles can contain links to other subreddits and or usernames. If these links contain underscores, Snudown parses these incorrectly if they contain underscores because the entire title is wrapped in a `[]()` link block, similar to the issue referenced in 7874382

This resolves the issue mentioned in [this post](https://reddit.com/r/bugs/comments/4rng30/_/).