LimitUploadSize: map maximum sizes by path and post data
Instead of one single maximum upload size, use a mapping to allow for endpoints to have a custom maximum post size.

The default is 500Kib as before, and entries are added to the map to be whitelisted. The path matches against the environ's PATH_INFO, stripped of a leading slash, and all keys selected must match the request's post data. A copy of the environ is sent to the map to keep webob from manipulating it and sending an environ with extra values to the downstream apps.