check_api_friend_oauth_scope: use dict.get instead of dict.__getitem__
VOneOf sets c.errors for us, so the user would be informed of his invalid option and get a 400 response. However, we can't use form.has_errrors because that would disallow the "enemy" type even when using the desktop site. At the same time, if a type other than those allowed and "enemy" are sent, we get a 500 because we are using `api_friend_scope_map`'s __getitem__. Switching to it's get function gives us the 400 error we want instead of the 500, and continues to allow the "enemy" type on desktop.