Use ARIA regions to group all elements belonging to a thing together.
reddit.com already uses role="main" to indicate where main content
of a page starts.  However, it is very unconfortable
to navigate from entry to entry.  This very small patch
adds a role="region" attribute around things, which tells screen
readers to indicate to the users that all elements inside that div
actually belong together.  This makes it easy to skip
to the next thing.