Use relative urls to the same (not the parent) directory in CSS
Relative urls to the parent directory on local installs (build/public) don't have a file to match, so Pylons takes over, and itself fails because the only thing it matches is a potential url (ex `droparrowgray.gif` being a domain name), and then redirects to a submit page to submit it.

Being relative to the same directory is what is needed on local installs, and what is supposed to be done on installs using a static domain. The only reason the real static domain still works is because all files are located in the root directory, and the parent directory of the root directory of a site is considered to be itself.

Two other changes were made to the spritesheet generator and on-the-fly less parser, because the images are located in the parent directory relative to the original `{name}.less/compact.css` file that it gets passed, so it needs to search one directory above now by itself.